+++ ChangeLog.md
@@ -3,9 +3,9 @@ ffdropenc Change Log
 
 ##2014-01-22 - v1.7##
 
- * **NEW**: make\_ffdropenc.sh - Automated build script for ffdropenc.app and distributable DMGs.  
+ * **NEW**: make-ffdropenc.sh - Automated build script for ffdropenc.app and distributable DMGs.  
  * **UPDATED**: Code reorganization to make further development easier.  
- * **UPDATED**: We no longer distribute ffmpeg or x264 binaries with the code. Use make\_ffdropenc.sh to compile your own.  
+ * **UPDATED**: We no longer distribute ffmpeg or x264 binaries through GitHub. Use make-ffdropenc.sh or ffmpeg-static to compile your own.  
  * **UPDATED**: MPG/MPEG extensions added to file filters.  
  * **NEW**: MKV (H.264/Audio Pass-thru) Preset. Useful for compressing DVDs for use by media servers.